+++ tests/Doubles/UrlPrintingCommand.php
@@ -4,6 +4,7 @@ namespace Tests\OriNette\Console\Doubles;
 
 use Nette\Http\IRequest;
 use Symfony\Component\Console\Command\Command;
+use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
 
@@ -26,6 +27,14 @@ final class UrlPrintingCommand extends Command
 	public static function getDefaultDescription(): string
 	{
 		return 'Print URL to output';
+	}
+
+	protected function configure(): void
+	{
+		// Ensures --ori-url is compatible with command arguments and options
+		$this->addArgument('required-arg', InputArgument::REQUIRED);
+		$this->addArgument('optional-arg', InputArgument::OPTIONAL);
+		$this->addOption('option');
 	}
 
 	protected function execute(InputInterface $input, OutputInterface $output): int